Moved upload and link-preview rendering to template.
Removed upload and link-preview rendering from the Message class, and
replaced it with wrapper code to expose the interesting parts of those
objects to the template.

Modified template and CSS to allow presenting the current attachment and
upload format, with all the HTML generation now done in templates instead
of in python.