Fix broken channel parsing
The previous change added a --channels flag that uses a default value of
'' when no argument is passed. It performs checking that it is a string
type before splitting, and then uses that in the list comprehension to
build all the channel names. As such, we would filter out all the
channel names, and cause a IndexError: list index out of range